+++ json_serializable/lib/src/json_key_helpers.dart
@@ -10,6 +10,8 @@ import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart' show alwaysThrows;
 import 'package:source_gen/source_gen.dart';
 
+import 'json_literal_generator.dart';
+
 @alwaysThrows
 T _throwUnsupported<T>(FieldElement element, String message) =>
     throw new InvalidGenerationSourceError(
@@ -43,11 +45,49 @@ JsonKeyWithConversion _from(FieldElement element) {
   var fromJsonName = _getFunctionName(obj, element, true);
   var toJsonName = _getFunctionName(obj, element, false);
 
+  Object _getLiteral(DartObject dartObject) {
+    if (dartObject.isNull) {
+      return null;
+    }
+
+    var reader = new ConstantReader(dartObject);
+
+    if (reader.isSymbol) {
+      _throwUnsupported(element,
+          'Values of type `Symbol` are not supported for `defaultValue`.');
+    } else if (reader.isType) {
+      _throwUnsupported(element,
+          'Values of type `Type` are not supported for `defaultValue`.');
+    } else if (!reader.isLiteral) {
+      _throwUnsupported(
+          element, 'The provided `defaultValue` is not a literal: $dartObject');
+    }
+
+    var literal = reader.literalValue;
+    if (literal is num || literal is String || literal is bool) {
+      return literal;
+    } else if (literal is List<DartObject>) {
+      return literal.map(_getLiteral).toList();
+    } else if (literal is Map<DartObject, DartObject>) {
+      return literal
+          .map((k, v) => new MapEntry(_getLiteral(k), _getLiteral(v)));
+    }
+    _throwUnsupported(
+        element, 'The provided value is not supported: $dartObject');
+  }
+
+  var defaultValueLiteral = _getLiteral(obj.getField('defaultValue'));
+
+  if (defaultValueLiteral != null) {
+    defaultValueLiteral = jsonLiteralAsDart(defaultValueLiteral, false);
+  }
+
   return new JsonKeyWithConversion._(
       obj.getField('name').toStringValue(),
       obj.getField('nullable').toBoolValue(),
       obj.getField('includeIfNull').toBoolValue(),
       obj.getField('ignore').toBoolValue(),
+      defaultValueLiteral,
       fromJsonName,
       toJsonName);
 }
@@ -69,12 +109,13 @@ class JsonKeyWithConversion extends JsonKey {
         super();
 
   JsonKeyWithConversion._(String name, bool nullable, bool includeIfNull,
-      bool ignore, this.fromJsonData, this.toJsonData)
+      bool ignore, Object defaultValue, this.fromJsonData, this.toJsonData)
       : super(
             name: name,
             nullable: nullable,
             includeIfNull: includeIfNull,
-            ignore: ignore);
+            ignore: ignore,
+            defaultValue: defaultValue);
 }
 
 ConvertData _getFunctionName(