+++ json_serializable/test/json_serializable_test.dart
@@ -603,4 +603,25 @@ abstract class _$SubTypeSerializerMixin {
 
     expect(output, expected);
   });
+
+  group('default values', () {
+    test('fail with symbols', () {
+      expectThrows(
+          'DefaultWithSymbol',
+          'Error with `@JsonKey` on `field`. '
+          'Values of type `Symbol` are not supported for `defaultValue`.');
+    });
+    test('fail with functions', () {
+      expectThrows(
+          'DefaultWithFunction',
+          'Error with `@JsonKey` on `field`. '
+          'The provided `defaultValue` is not a literal: () → int (_function)');
+    });
+    test('fail with type', () {
+      expectThrows(
+          'DefaultWithType',
+          'Error with `@JsonKey` on `field`. '
+          'Values of type `Type` are not supported for `defaultValue`.');
+    });
+  });
 }