+++ json_serializable/test/src/default_value_input.dart
@@ -1,0 +1,33 @@
+// Copyright (c) 2018, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+part of 'json_serializable_test_input.dart';
+
+@JsonSerializable()
+class DefaultWithSymbol {
+  @JsonKey(defaultValue: #symbol)
+  Object field;
+
+  DefaultWithSymbol();
+}
+
+int _function() => 42;
+
+@JsonSerializable()
+class DefaultWithFunction {
+  @JsonKey(defaultValue: _function)
+  Object field;
+
+  DefaultWithFunction();
+}
+
+@JsonSerializable()
+class DefaultWithType {
+  @JsonKey(defaultValue: Object)
+  Object field;
+
+  DefaultWithType();
+}
+
+// TODO: const, non-literal (const Duration, etc)