+++ json_serializable/test/src/json_serializable_test_input.dart
@@ -5,6 +5,7 @@
 //ignore_for_file: avoid_unused_constructor_parameters, prefer_initializing_formals
 import 'package:json_annotation/json_annotation.dart';
 
+part 'default_value_input.dart';
 part 'generic_test_input.dart';
 part 'to_from_json_test_input.dart';